feat: rename aipspansql to spanordering
For naming parity between the following packages:

* `spanordering` (to transpile ordering.OrderBy to Spanner SQL)
* `spanfiltering` (to transpile filtering.Filter to Spanner SQL)

BREAKING CHANGE: aipspansql.Order package has been renamed to
                 spanordering.TranspileOrderBy.
