+++ doc.go
@@ -1,5 +1,5 @@
-// Package aipspanner provides primitives for implementing Spanner persistence
+// Package spanneraip is a collection of packages for implementing Spanner persistence
 // for resource-oriented APIs.
 //
 // See: https://aip.dev
-package aip
+package spanneraip
