+++ spanordering/doc.go
@@ -1,0 +1,4 @@
+// Package spanordering provides primitives for transpiling AIP ordering directives to Spanner SQL.
+//
+// See: https://google.aip.dev/132#ordering
+package spanordering
